test_requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile test_requirements.in --output-file test_requirements.txt --upgrade
#
adal==1.2.1
apipkg==1.5               # via execnet
appdirs==1.4.3            # via black
asn1crypto==0.24.0        # via cryptography
asynctest==0.12.4
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via black, jsonschema, pytest
backcall==0.1.0           # via ipython
black==19.3b0
bleach==3.1.0             # via nbconvert
certifi==2019.3.9         # via requests
cffi==1.12.3              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via black
cryptography==2.6.1       # via adal
decorator==4.4.0          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
docker-pycreds==0.4.0     # via docker
docker==3.7.0
entrypoints==0.3          # via nbconvert
execnet==1.6.0            # via pytest-xdist
idna==2.8                 # via requests
ipykernel==5.1.0          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.5.0            # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.4.2         # via jupyter
jedi==0.13.3              # via ipython
jinja2==2.10.1            # via nbconvert, notebook
jsonschema==3.0.1         # via nbformat
jupyter-client==5.2.4     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==6.0.0    # via jupyter
jupyter-core==4.4.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
markupsafe==1.1.1         # via jinja2
mistune==0.8.4            # via nbconvert
more-itertools==7.0.0     # via pytest
mypy-extensions==0.4.1    # via mypy
mypy==0.701               # via pytest-mypy
nbconvert==5.4.0
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==5.7.8
pandocfilters==1.4.2      # via nbconvert
parso==0.4.0              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.9.0             # via pytest
prometheus-client==0.6.0  # via notebook
prompt-toolkit==2.0.9     # via ipython, jupyter-console
ptyprocess==0.6.0         # via pexpect, terminado
py==1.8.0                 # via pytest
pycparser==2.19           # via cffi
pyflakes==2.1.1           # via pytest-flakes
pygments==2.3.1           # via ipython, jupyter-console, nbconvert, qtconsole
pyjwt==1.7.1              # via adal
pyrsistent==0.15.1        # via jsonschema
pytest-asyncio==0.10.0
pytest-flakes==4.0.0
pytest-forked==1.0.2      # via pytest-xdist
pytest-mypy==0.3.2
pytest-timeout==1.3.3
pytest-xdist==1.26.1
pytest==4.0.0
python-dateutil==2.8.0    # via adal, jupyter-client
pyzmq==18.0.1             # via jupyter-client, notebook
qtconsole==4.4.3          # via jupyter
requests==2.21.0          # via adal, docker, responses
responses==0.10.5
ruamel.yaml==0.15.76
send2trash==1.5.0         # via notebook
six==1.12.0               # via bleach, cryptography, docker, docker-pycreds, jsonschema, prompt-toolkit, pyrsistent, pytest, pytest-xdist, python-dateutil, responses, traitlets, websocket-client
terminado==0.8.2          # via notebook
testpath==0.4.2           # via nbconvert
toml==0.10.0              # via black
tornado==5.1.1
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
typed-ast==1.3.5          # via mypy
urllib3==1.24.2
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via bleach
websocket-client==0.56.0  # via docker
widgetsnbextension==3.4.2  # via ipywidgets
